Ct updated to allow repeats in the atom selection
Previously, the bonds were defined with two selections of atoms, which meant that two bonds having a common atom could not be used (since there is no way to have two copies of an atom in the selection). Now, we allow an index to refer to atoms in the selection. If no duplicates, then index can run from 0 to # of atoms. (mol.sel1in, mol.sel2in now used)